+++ tests/table/unit/query.interpreter.unit.test.ts
@@ -32,6 +32,7 @@ describe("Query Interpreter", () => {
       double: -123.01,
       bool: true,
       date: "2020-01-01T00:00:00.000Z",
+      microsecondDate: "2023-01-01T00:00:00.000000Z",
       guid: Buffer.from("00000000-0000-0000-0000-000000000000").toString("base64"),
       guidLegacy: "00000000-0000-0000-0000-000000000000",
       binary: Buffer.from("binaryData").toString("base64"),
@@ -240,6 +241,129 @@ describe("Query Interpreter", () => {
         name: "GUID inequality (legacy)",
         originalQuery: "guidLegacy ne guid'22222222-2222-2222-2222-222222222222'",
         expectedResult: true
+      }
+    ])
+
+    runTestCases("DateTimes", referenceEntity, [
+      {
+        name: "DateTime equality",
+        originalQuery: "date eq datetime'2020-01-01T00:00:00.000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime equality (doesn't match)",
+        originalQuery: "date eq datetime'2020-01-01T00:00:01.000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime equality (microseconds)",
+        originalQuery: "microsecondDate eq datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime equality (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate eq datetime'2023-01-01T00:00:00.001000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime inequality",
+        originalQuery: "date ne datetime'2020-01-01T00:00:01.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime inequality (doesn't match)",
+        originalQuery: "date ne datetime'2020-01-01T00:00:00.000000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime inequality (microseconds)",
+        originalQuery: "microsecondDate ne datetime'2023-01-01T00:00:01.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime inequality (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate ne datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime greater than",
+        originalQuery: "date gt datetime'2019-12-31T23:59:59.999999Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime greater than (doesn't match)",
+        originalQuery: "date gt datetime'2020-01-01T00:00:00.000000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime greater than (microseconds)",
+        originalQuery: "microsecondDate gt datetime'2022-12-31T23:59:59.999999Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime greater than (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate gt datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime greater than or equal",
+        originalQuery: "date ge datetime'2020-01-01T00:00:00.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime greater than or equal (doesn't match)",
+        originalQuery: "date ge datetime'2020-01-01T00:00:00.001000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime greater than or equal (microseconds)",
+        originalQuery: "microsecondDate ge datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime greater than or equal (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate ge datetime'2023-01-01T00:00:00.001000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime less than",
+        originalQuery: "date lt datetime'2020-01-01T00:00:00.001Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime less than (doesn't match)",
+        originalQuery: "date lt datetime'2020-01-01T00:00:00.000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime less than (microseconds)",
+        originalQuery: "microsecondDate lt datetime'2023-01-01T00:00:00.001000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime less than (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate lt datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime less than or equal",
+        originalQuery: "date le datetime'2020-01-01T00:00:00.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime less than or equal (doesn't match)",
+        originalQuery: "date le datetime'2019-12-31T23:59:59.999999Z'",
+        expectedResult: false
+      },
+      {
+        name: "DateTime less than or equal (microseconds)",
+        originalQuery: "microsecondDate le datetime'2023-01-01T00:00:00.000000Z'",
+        expectedResult: true
+      },
+      {
+        name: "DateTime less than or equal (microseconds) (doesn't match)",
+        originalQuery: "microsecondDate le datetime'2022-12-31T23:59:59.999999Z'",
+        expectedResult: false
       }
     ])
   })